Move config consumer folder
