--- pkg/kafka/config/producer/.nocover
@@ -1,1 +1,0 @@
-requires connection to Kafka
